--- geode_geocoding/admin.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-from django.contrib import admin
-
-# Register your models here.
