--- geode_geocoding/models.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from django.db import models
